backend/requirements.txt
# Core
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.6.0

# Authentication
djangorestframework-simplejwt==5.4.0

# Database
psycopg[binary]==3.2.4

# Redis & Caching
django-redis==5.4.0
redis==5.2.1

# Real-time (WebSocket - Phase 6)
# channels==4.2.0
# channels-redis==4.2.1

# Environment
python-decouple==3.8

# Development
django-debug-toolbar==4.4.6
django-extensions==3.2.3

# Code Quality
ruff==0.9.6
